server/requirements.txt
Django>=2.0,<3.0
djangorestframework==3.9.2